Made it possible to load individual images of colonies from anywhere
