+++ scanomatic/qc/analysis_results.py
@@ -4,16 +4,16 @@ import os
 import re
 import types
 
+from matplotlib import pyplot as plt
+
 from scanomatic.generics.maths import mid50_mean
 from scanomatic.io.image_data import ImageData
+from scanomatic.io.image_loading import load_colony_images_for_animation
 from scanomatic.io.logger import Logger
 from scanomatic.io.movie_writer import MovieWriter
 from scanomatic.io.paths import Paths
-from scanomatic.models.factories.compile_project_factory import CompileImageAnalysisFactory
 
-from matplotlib import pyplot as plt
 # This import is used in 3D plotting just not explicitly stupid matplotlib
-from mpl_toolkits.mplot3d import Axes3D
 
 _pattern = re.compile(r".*_([0-9]+)_[0-9]+_[0-9]+_[0-9]+\..*")
 _logger = Logger("Phenotype Results QC")
@@ -81,122 +81,6 @@ def plot_growth_curve(growth_data, position, ax=None, save_target=None):
 def set_axvspan_width(polygon, width):
 
     polygon.xy[2:4, 0] = width
-
-
-def load_colony_images_for_animation(analysis_directory, position, project_compilation=None, positioning="one-time"):
-    """
-
-    :param analysis_directory: path to analysis directory
-    :type  analysis_directory: str
-    :param position: list/tuple of colony to extract. (Plate, Row, Colum) Note that all positions should be
-    enumerated from 1. because confusion!
-    :type position: [int]
-    :param project_compilation: Path to the associated compilation file, inferred if not submitted
-    :type project_compilation: str
-    :param positioning: Type of positioning to simulate. Default is "one-time", which uses the gridding image
-    positioning all through.Use "detected" for the position actually detected.
-    :type positioning: str
-    :return: First array is a 1D time-vector, second array is a 3D image sequence vector where the last dimension
-    is time, the third array is the 2D plate slice of the last image.
-    :rtype : numpy.ndarray, numpy.ndarray, numpy.ndarray
-    """
-
-    # TODO: Verify that it is the correct image that is displayed
-
-    def _bound(bounds, a, b):
-
-        def bounds_check(bound, val):
-
-            if 0 <= val < bound:
-                return val
-            elif val < 0:
-                return 0
-            else:
-                return bound - 1
-
-        return ((bounds_check(bounds[0], a[0]),
-                 bounds_check(bounds[0], a[1])),
-                (bounds_check(bounds[1], b[0]),
-                 bounds_check(bounds[1], b[1])))
-
-    def slice_im(plate, colony_position, colony_size):
-
-        lbound = colony_position - np.floor(colony_size / 2)
-        ubound = colony_position + np.ceil(colony_size / 2)
-        if (ubound - lbound != colony_size).any():
-            ubound += colony_size - (ubound - lbound)
-
-        return plate[lbound[0]: ubound[0], lbound[1]:ubound[1]]
-
-    assert positioning in ('one-time', 'detected'), "Unknown positioning argument"
-
-    analysis_directory = os.path.abspath(analysis_directory)
-
-    if not project_compilation:
-
-        experiment_directory = os.sep.join(analysis_directory.split(os.sep)[:-1])
-        experiment_name = experiment_directory.split(os.sep)[-1]
-
-        project_compilation = os.path.join(experiment_directory,
-                                           Paths().project_compilation_pattern.format(experiment_name))
-
-        if not os.path.isfile(project_compilation):
-
-            candidates = glob.glob(os.path.join(
-                experiment_directory, Paths().project_compilation_pattern.format("*")))
-
-            if not candidates:
-                _logger.error("Could not find any project.compilation file in '{0}'".format(experiment_directory))
-                raise ValueError()
-            elif len(candidates) != 1:
-                _logger.error(
-                    "Found several project.compilation files in '{0}', unsure which to use.".format(
-                            experiment_directory) +
-                    "Either remove one of {0} or specify compilation-file in function call".format(candidates))
-                raise ValueError()
-
-            project_compilation = candidates[0]
-    else:
-        experiment_directory = os.path.dirname(project_compilation)
-
-    # grids number +1
-    grid = np.load(os.path.join(analysis_directory, Paths().grid_pattern.format(position[0] + 1)))
-    grid_size = np.load(os.path.join(analysis_directory, Paths().grid_size_pattern.format((position[0] + 1))))
-
-    compilation_results = CompileImageAnalysisFactory.serializer.load(project_compilation)
-    compilation_results = sorted(compilation_results, key=lambda e: e.image.index)
-
-    times = np.array(tuple(entry.image.time_stamp for entry in compilation_results))
-    images = np.zeros(tuple(grid_size) + times.shape, dtype=np.uint16)
-    im = None
-    ref_plate_model = compilation_results[-1].fixture.plates
-
-    for i, entry in enumerate(compilation_results):
-        try:
-            im = plt.imread(entry.image.path)
-        except IOError:
-            im = plt.imread(os.path.join(experiment_directory, os.path.basename(entry.image.path)))
-
-        if positioning == 'one-time':
-            plate_model = ref_plate_model[position[0]]
-        elif positioning == 'detected':
-            plate_model = entry.fixture.plates[position[0]]
-        else:
-            raise ValueError("Positioning can't be '{0}'".format(positioning))
-
-        x = sorted((plate_model.x1, plate_model.x2))
-        y = sorted((plate_model.y1, plate_model.y2))
-
-        y, x = _bound(im.shape, y, x)
-
-        # As gridding is done on plates as seen in the scanner while plate positioning is done on plates
-        # as seen by the scanner the inverse direction of the short dimension is needed and needed after
-        # slicing out the plate
-        im = im[y[0]: y[1], x[0]: x[1]][:, ::-1]
-
-        images[..., i] = slice_im(im, grid[:, grid.shape[1] - position[2] - 1, position[1]], grid_size)
-
-    return times, images, im
 
 
 def animate_colony_growth(save_target, position, analysis_folder, fps=12, project_compilation=None, fig=None,
